inject component
